fix: adds authentication when deleting intellectual property and works, validating the password before the operation

+++ src/lib/api/controllers/intellectuals/deleteIntellectualsService.ts
@@ -1,8 +1,15 @@
 import { Message } from "@/lib/db/models/Message";
 import { DeleteIntellectualsService } from "../../services/intellectuals/deleteIntellectualsService";
+import { Auth } from "../../auth/authDelete";
 
 export class DeleteIntellectualsController {
-  async deleteController(id: string) {
+  async deleteController(id: string, pass: string) {
+    const auth = Auth(pass);
+    
+    if (!auth) {
+      return new Message(false, "the password is incorrect", null);
+    }
+
     const service = new DeleteIntellectualsService();
 
     const response = await service.deleteService(id);
